Delete no record found
There is no point in returning no record found as error any more.
Server process a 200 response since it a valid input.
Other issue to take into account is that looking for an user should
report an error message, because app can't validate if username is
in valid format. As long as it's a string is valid. Leaving a no
user found error with a 404 makes sense.