Move update method to model
Instead of creating an additional structure with it own fields, take
the advantage of an existing one. Since the app is already loading in
memory the whole entity, it's better to use it already. As an added
benefit, all model fields are already in the struct, no need to look
around.